requirements.txt
Django==2.1.2
django-debug-toolbar==1.10.1
django-environ==0.4.5
gunicorn==19.9.0
psycopg2-binary==2.7.5
pytz==2018.5
sqlparse==0.2.4
whitenoise==4.1
